requirements.txt
PyVISA==1.5
mechanize==0.2.5
BeautifulSoup==3.2.1
pyserial==2.7
